requirements.txt
fastapi
uvicorn
celery
redis
crewai
langchain-google-genai
requests
python-dotenv
pydantic
gevent
pytest 
pytest-asyncio 
httpx